% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_journal.R
\name{set_journal}
\alias{set_journal}
\title{Set Default Journal}
\usage{
set_journal(journal)
}
\arguments{
\item{journal}{Character. Name of the journal (e.g., "Nature").}
}
\description{
Sets a global option for the target journal. This prevents you from needing
to specify the \code{journal} argument in every call to \code{publish_fig()}.
}
