% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_paths.R
\name{set_paths}
\alias{set_paths}
\title{Set Default Output Paths for Figures}
\usage{
set_paths(..., create = FALSE)
}
\arguments{
\item{...}{Character. One or more directory paths.}

\item{create}{Logical. If \code{TRUE}, create any directories that do not
already exist. Default \code{FALSE}.}
}
\value{
Invisibly returns the registered paths.
}
\description{
Registers one or more output directories that \code{publish_fig()} will
automatically save to. When paths are registered, passing a bare filename
(i.e. \code{figureout.name}) to \code{publish_fig()} will save the figure to
every registered directory, eliminating the need for manual path construction.
}
\details{
Paths are stored in the R option \code{figureout_paths} for the duration of the
session. Call \code{set_paths()} with no arguments to clear all registered
paths.
}
\examples{
\dontrun{
set_paths("report/assets", "figures")
# publish_fig() will now save to both directories automatically
}

}
\seealso{
\code{\link{publish_fig}}, \code{\link{set_journal}}
}
