% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_journals.R
\name{show_journals}
\alias{show_journals}
\title{List Available Journals}
\usage{
show_journals()
}
\value{
A data frame with one column, \code{Journals}, containing sorted journal names.
}
\description{
Returns a data frame of journal names for which default figure dimensions are available.
}
