% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_edge_effect.R
\name{theme_edge}
\alias{theme_edge}
\title{Initialize plotting environment}
\usage{
theme_edge()
}
\value{
A named list of plotting defaults.
}
\description{
Defines palettes, labels, line widths, and shading color used for figures.
}
